Refactor: 클라이언트 클래스 추가
