+++ src/types/ClientConfig.ts
@@ -1,0 +1,4 @@
+export interface ClientConfig {
+  clientId: string;
+  clientSecret: string;
+}
